auth for websocket
