+++ index.js
@@ -3,20 +3,7 @@
 
 require("dotenv").config();
 
-const OktaJwtVerifier = require("@okta/jwt-verifier");
-const { access } = require("fs");
-var https = require("https");
-const jsonWebToken = require("jsonwebtoken");
-
-/******************************************************/
-
-const oktaJwtVerifier = new OktaJwtVerifier({
-  issuer: process.env.ISSUER, // required
-  clientId: process.env.CLIENT_ID, // required
-  assertClaims: {
-    aud: process.env.AUDIENCE,
-  },
-});
+const VerifyToken = require("./verify-token.js");
 
 const AuthPolicy = require("./auth-policy");
 
@@ -55,8 +42,6 @@ const allowAccess = (event, email) => {
   return policy;
 };
 
-/******************************************************/
-
 exports.handler = function (event, context) {
   var arr = event.authorizationToken.split(" ");
 
@@ -64,70 +49,5 @@ exports.handler = function (event, context) {
 
   console.log("Access token: " + accessToken);
 
-  oktaJwtVerifier
-    .verifyAccessToken(accessToken, process.env.AUDIENCE)
-    .then((jwt) => {
-      // the token is valid (per definition of 'valid' above)
-      console.log("okta request principal: " + JSON.stringify(jwt.claims));
-
-      var policy = allowAccess(event, jwt.claims.sub);
-
-      return context.succeed(policy.build({ groups: "Ad-Who2-Users" }));
-      // return context.succeed(policy.build({ groups: jwt.claims.groups.join(',') }));
-    })
-    .catch((err) => {
-      console.log(err);
-      if (err) {
-        var decoded = jsonWebToken.decode(accessToken);
-
-        if (
-          !decoded ||
-          ![process.env.AAD_APPLICATION_ID, process.env.AAD_CN_APPLICATION_ID]
-            .filter((value) => !!value)
-            .includes(decoded.appid)
-        ) {
-          console.error("Decoded token is " + JSON.stringify(decoded));
-          return context.fail(
-            "Unauthorized due to invalid aad application id and failed Okta auth"
-          );
-        }
-        var params = {
-          host: "graph.microsoft.com",
-          path: "/v1.0/me",
-          port: 443,
-          headers: { Authorization: `Bearer ${accessToken}` },
-        };
-
-        https.get(params, (response) => {
-          let chunksOfData = [];
-
-          response.on("data", (fragments) => {
-            chunksOfData.push(fragments);
-          });
-
-          response.on("end", () => {
-            let responseBody = JSON.parse(
-              Buffer.concat(chunksOfData).toString()
-            );
-            if (responseBody.userPrincipalName != decoded.upn) {
-              console.error(
-                `${responseBody.userPrincipalName} do not match with ${decoded.upn}`
-              );
-              console.error(responseBody);
-              return context.fail("Unauthorized");
-            }
-
-            var policy = allowAccess(event, responseBody.userPrincipalName);
-
-            return context.succeed(policy.build({ groups: "Ad-Who2-Users" }));
-          });
-
-          response.on("error", (error) => {
-            console.error(error);
-            console.error("Decoded token is " + JSON.stringify(decoded));
-            return context.fail("Unauthorized due to AAD auth failed");
-          });
-        });
-      }
-    });
+  return VerifyToken.verifyAccessToken(accessToken, event, context, allowAccess);
 };
